+++ src/term/methods/tag/canBe.js
@@ -9,14 +9,14 @@ const canBe = function(term, tag) {
     return true;
   }
   //loop through tag's contradictory tags
-  let enemies = tagset[tag].enemy || [];
+  let enemies = tagset[tag].isNot || [];
   for (let i = 0; i < enemies.length; i++) {
     if (term.tags[enemies[i]] === true) {
       return false;
     }
   }
-  if (tagset[tag].is !== undefined) {
-    return canBe(term, tagset[tag].is); //recursive
+  if (tagset[tag].isA !== undefined) {
+    return canBe(term, tagset[tag].isA); //recursive
   }
   return true;
 };