+++ src/term/methods/tag/setTag.js
@@ -19,15 +19,15 @@ const putTag = (term, tag, reason) => {
   //extra logic per-each POS
   if (tagset[tag]) {
     //drop any conflicting tags
-    let enemies = tagset[tag].enemy;
+    let enemies = tagset[tag].isNot || [];
     for (let i = 0; i < enemies.length; i++) {
       if (term.tags[enemies[i]] === true) {
         unTag(term, enemies[i], reason);
       }
     }
     //apply implicit tags
-    if (tagset[tag].is) {
-      let doAlso = tagset[tag].is;
+    if (tagset[tag].isA) {
+      let doAlso = tagset[tag].isA;
       if (term.tags[doAlso] !== true) {
         putTag(term, doAlso, ' --> ' + tag); //recursive
       }
@@ -36,13 +36,13 @@ const putTag = (term, tag, reason) => {
 };
 
 //give term this tag
-const wrap = function (term, tag, reason) {
+const wrap = function(term, tag, reason) {
   if (!term || !tag) {
     return;
   }
   //handle multiple tags
   if (fns.isArray(tag)) {
-    tag.forEach((t) => putTag(term, t, reason)); //recursive
+    tag.forEach(t => putTag(term, t, reason)); //recursive
     return;
   }
   putTag(term, tag, reason);